# public/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/aircraft/fetch": {
      "get": {
        "operationId": "AircraftController_fetch",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Aircraft"
        ]
      }
    },
    "/weather-analytics/basic-stats": {
      "get": {
        "operationId": "WeatheranalyticsController_getStats",
        "parameters": [
          {
            "name": "interval",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "airportId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weatheranalytics"
        ]
      }
    },
    "/weather-analytics/timeseries": {
      "get": {
        "operationId": "WeatheranalyticsController_getTimeSeries",
        "parameters": [
          {
            "name": "interval",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "airportId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weatheranalytics"
        ]
      }
    },
    "/weather-analytics/predict-next-day{airportId}": {
      "get": {
        "operationId": "WeatheranalyticsController_getNextDayPrediction",
        "parameters": [
          {
            "name": "airportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weatheranalytics"
        ]
      }
    },
    "/locations/airports": {
      "get": {
        "operationId": "LocationsController_getAllAirportLocations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/bmkg-radars": {
      "get": {
        "operationId": "LocationsController_getAllBMKGRadarSites",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Locations"
        ]
      }
    }
  },
  "info": {
    "title": "MET Flight Radar API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {}
  }
}